+++ tests/compile_fail/invalid_align_attr.rs
@@ -4,12 +4,12 @@ fn main() {}
 
 #[derive(ShaderType)]
 struct Test {
-    #[align]
+    #[shader_align]
     a: u32,
-    #[align()]
+    #[shader_align()]
     b: u32,
-    #[align(invalid)]
+    #[shader_align(invalid)]
     c: u32,
-    #[align(3)]
+    #[shader_align(3)]
     d: u32,
 }