+++ tests/pass/attributes.rs
@@ -4,7 +4,7 @@ fn main() {}
 
 #[derive(ShaderType)]
 struct TestAttributes {
-    #[align(16)]
+    #[shader_align(16)]
     a: u32,
     #[size(8)]
     b: u32,
@@ -14,7 +14,7 @@ struct TestAttributes {
 struct TestRtArray {
     #[size(8)]
     a: u32,
-    #[align(16)]
+    #[shader_align(16)]
     #[size(runtime)]
     b: Vec<u32>,
 }